updated project structure to support both IDE and CLI
