+++ README.md
@@ -1,0 +1,41 @@
+# Python Machine Learning
+
+> Machine-learning practice includes classifier, and etc.
+
+
+
+<br /><a name="contents"></a>
+## Contents
+
+* [Classifier](./classifier)
+* [Project Setup](#setup)
+* [Start](#start)
+
+
+<br /><a name="setup"></a>
+## Setup
+
+### Cleaning-up
+
+  ```
+  make clean
+  ```
+
+
+
+<br /><a name="start"></a>
+## Start
+
+### Classifier
+
+  * Run training:
+
+    ```
+    make ctr  # or run `(cd ml/classifier; make training)`
+    ```
+
+  * Run prediction:
+
+    ```
+    make cpr  # or run `(cd ml/classifier; make prediction)`
+    ```
