--- classifier/Makefile
@@ -1,5 +1,0 @@
-.PHONY: clean training
-
-
-training tr:
-	python training.py
