--- classifier/utils.py
@@ -1,41 +1,0 @@
-"""
-
-"""
-import numpy as np
-from matplotlib import pyplot as plt
-import h5py
-
-
-def load_data():
-    train_dataset = h5py.File('datasets/train_catvnoncat.h5', "r")
-    #print(type(train_dataset))
-    train_set_x_orig = np.array(train_dataset["train_set_x"][:])  # your train set features
-    train_set_y_orig = np.array(train_dataset["train_set_y"][:])  # your train set labels
-
-    test_dataset = h5py.File('datasets/test_catvnoncat.h5', "r")
-    test_set_x_orig = np.array(test_dataset["test_set_x"][:])  # your test set features
-    test_set_y_orig = np.array(test_dataset["test_set_y"][:])  # your test set labels
-
-    classes = np.array(test_dataset["list_classes"][:])  # the list of classes
-
-    train_set_y_orig = train_set_y_orig.reshape((1, train_set_y_orig.shape[0]))
-    test_set_y_orig = test_set_y_orig.reshape((1, test_set_y_orig.shape[0]))
-
-    return train_set_x_orig, train_set_y_orig, test_set_x_orig, test_set_y_orig, classes
-
-
-def print_mislabeled_images(classes, X, y, p):
-
-    a = p + y
-    mislabeled_indices = np.asarray(np.where(a == 1))
-    plt.rcParams['figure.figsize'] = (40.0, 40.0)  # set default size of plots
-    num_images = len(mislabeled_indices[0])
-    for i in range(num_images):
-        index = mislabeled_indices[1][i]
-
-        plt.subplot(2, num_images, i + 1)
-        plt.imshow(X[:, index].reshape(64, 64, 3), interpolation='nearest')
-        plt.axis('off')
-        plt.title(
-            "Prediction: " + classes[int(p[0, index])].decode("utf-8") + " \n Class: " + classes[y[0, index]].decode(
-                "utf-8"))
