+++ ml/requirements.txt
@@ -1,0 +1,10 @@
+cycler==0.10.0
+h5py==2.9.0
+kiwisolver==1.0.1
+matplotlib==3.0.2
+numpy==1.15.4
+Pillow==5.4.1
+pyparsing==2.3.0
+python-dateutil==2.7.5
+scipy==1.2.0
+six==1.12.0
