+++ ml/setup.cfg
@@ -1,0 +1,30 @@
+[coverage:run]
+omit =
+  **/tests/*
+  **/setup.py
+  setup.py
+  .venv/*
+
+[nosetests]
+# nosetests can only run with this setup.cfg in the same path
+
+cover-package=ml-practice
+cover-html=1
+cover-min-percentage=96
+detailed-errors=1
+nocapture=1
+tissue-color=1
+# tissue ignore E501 (line /too long) since no option to set max-line-length
+tissue-ignore=E501
+tissue-package=ml-practice
+with-tissue=1
+with-coverage=1
+verbosity=3
+
+[tool:pytest]
+addopts = --cov=. --cov-report=term --cov-report=html --cov-fail-under=95
+norecursedirs =
+  tests/*
+  .venv/*
+pep8maxlinelength = 120
+# pep8ignore = ALL
